Use 32-bit integers for line numbers
This saves 16% time, 18% memory and 1% instruction counts from the
sample files in #293. These are large files, treated as textual diffs,
so reducing the storage per-line is significant. Files with structural
diffing, such as sample_files/slow_before.rs, are unchanged.

Fixes #293
Closes #297

+++ src/diff/dijkstra.rs
@@ -223,7 +223,7 @@ mod tests {
     use itertools::Itertools;
     use typed_arena::Arena;
 
-    fn pos_helper(line: usize) -> Vec<SingleLineSpan> {
+    fn pos_helper(line: u32) -> Vec<SingleLineSpan> {
         vec![SingleLineSpan {
             line: line.into(),
             start_col: 0,
@@ -231,7 +231,7 @@ mod tests {
         }]
     }
 
-    fn col_helper(line: usize, col: u32) -> Vec<SingleLineSpan> {
+    fn col_helper(line: u32, col: u32) -> Vec<SingleLineSpan> {
         vec![SingleLineSpan {
             line: line.into(),
             start_col: col,
